style(v3.0.2): L1 fmt + clippy clean + websocket send bugfix (PR-2)
CONSENSUS-SAFE: genesis hash unchanged (7543004c...2728), KAT 162 pass.

- cosmic-harmony: clippy --fix (manual_swap, needless_borrow) + scoped
  #![allow(needless_range_loop)] where loops mirror the GPU kernel byte-for-byte;
  dead_code allow on scalar reference fns; # Safety docs on NEON unsafe fns.
- core: from_ref, iter().cloned(), match->let, &Path, redundant wildcard,
  dead_code/large_enum_variant allows.
  BUGFIX websocket.rs: ClientSession::send was async with a sync body, so the
  synchronous broadcast() dropped the returned future and never sent anything.
  Made send() a plain fn so broadcasts actually fire.
- pool/miner: clippy --fix + scoped allows for feature-gated GPU/UI scaffolding;
  banner cfg target_feature crypto->aes; removed unused rayon import.
- native-ffi: unused_mut allow (feature-conditional).

Verified: cargo fmt --all --check = 0, clippy --workspace -D warnings = 0,
cosmic-harmony/core/pool/miner tests pass.

+++ V3/L1/core/src/wallet.rs
@@ -331,9 +331,9 @@ pub fn generate_account_tx_id(from: &str, to: &str, amount: u64, nonce: u64) -> 
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
         .as_nanos();
-    bytes[..16].copy_from_slice(&(ts as u128).to_le_bytes());
-    bytes[16..24].copy_from_slice(&(amount as u64).to_le_bytes());
-    bytes[24..32].copy_from_slice(&(nonce as u64).to_le_bytes());
+    bytes[..16].copy_from_slice(&ts.to_le_bytes());
+    bytes[16..24].copy_from_slice(&amount.to_le_bytes());
+    bytes[24..32].copy_from_slice(&nonce.to_le_bytes());
     // XOR-in sender and recipient for uniqueness
     for (i, b) in from.bytes().chain(to.bytes()).enumerate() {
         bytes[i % 32] ^= b;
